api completed and categories and posts fetched from db to home page

+++ app/api/posts/[id]/route.ts
@@ -1,4 +1,6 @@
+import { authOptions } from "@/app/utilis/authoptions";
 import prisma from "@/lib/prismadb";
+import { getServerSession } from "next-auth/next";
 import { NextResponse } from "next/server";
 
 export async function GET(req: Request, {params}:{params: {id:string}}) {
@@ -15,6 +17,10 @@ export async function GET(req: Request, {params}:{params: {id:string}}) {
 
 
 export async function PUT(req: Request, {params}:{params: {id:string}}) {
+    const session = await getServerSession(authOptions);
+        if(!session){
+            return NextResponse.json({error: "Not authenticated"}, {status: 401})
+        }
     const { title, content, links, selectedCategory, imageUrl, publicId} = await req.json();
     const id = params.id;
     try {
@@ -33,6 +39,11 @@ export async function PUT(req: Request, {params}:{params: {id:string}}) {
 }
 
 export async function DELETE(req: Request, {params}:{params: {id:string}}){
+    const session = await getServerSession(authOptions);
+    if(!session){
+        return NextResponse.json({error: "Not authenticated"}, {status: 401})
+    }
+    
     const id = params.id 
     try{
         const post = await prisma.post.delete({where: {id}});
